refactor: move scan status to v1 package
