+++ internal/scan/v1/scan.go
